terminei de mudar dados, endereço, cartoes, perfil e home, falta o resto
