+++ app/src/main/AndroidManifest.xml
@@ -12,10 +12,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Login">
         <activity
-            android:name=".DadosPessoais"
+            android:name=".DadosPessoaisFragment"
             android:exported="false" />
         <activity
-            android:name=".MeusEnderecos"
+            android:name=".MeusEnderecosFragment"
             android:exported="false" />
 
         <activity
